lambda/requirements.txt
boto3==1.9.216
ask-sdk-core==1.11.0
mysql-connector-python==8.0.11